requirements.txt
aiohttp
aiosignal==1.3.1
altair==5.5.0
aniso8601==8.0.0
annotated-types==0.6.0
anthropic==0.36.2
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arxiv==2.1.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
boto3==1.34.162
botocore==1.34.162
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
# chromadb==0.5.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.4
DateTime==5.2
debugpy==1.8.2
python-docx
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
Django==4.2.7
django-crispy-forms==2.1
djangorestframework==3.14.0
dnspython==2.6.1
email_validator==2.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
environs==9.5.0
executing==2.0.1
# faiss-cpu==1.8.0
fastapi
fastapi-cli
feedparser==6.0.10
filelock==3.13.1
Flask
Flask-JWT-Extended
Flask-MySQLdb
Flask-RESTful
Flask-SQLAlchemy
flatbuffers==24.3.25
free_proxy==1.1.1
frozenlist==1.4.1
fsspec==2024.2.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.21.0
google-auth==2.30.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.1
graphviz==0.20.1
greenlet==3.1.1
groq==0.11.0
grpcio
grpcio-status
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.3
humanfriendly==10.0
hypercorn==0.17.3
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
iso8601==1.1.0
itsdangerous==1.1.0
jedi==0.19.1
Jinja2==2.11.2
jiter==0.6.1
jmespath==1.0.1
joblib==1.3.2
jsonify==0.5
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kubernetes==30.1.0
langchain
# langchain-anthropic==0.1.13
# langchain-aws==0.1.6
# langchain-chroma==0.1.1
# langchain-community==0.0.38
langchain-core
langchain-google-genai
# langchain-groq==0.1.3
# langchain-openai==0.1.1
# langchain-text-splitters==0.0.2
# langcodes==3.3.0
# langsmith==0.1.77
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==1.1.1
marshmallow==3.20.2
matplotlib-inline==0.1.7
mdurl==0.1.2
messagebox==0.1.0
# milvus==2.2.16
minify_html==0.15.0
minio==7.2.4
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
mysqlclient==2.2.4
narwhals==1.18.3
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.2.1
onnxruntime==1.18.0
openai==1.34.0
opencv-python==4.10.0.84
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
# opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.0.3
parso==0.8.4
Pillow==10.0.0
platformdirs==4.2.0
playwright==1.48.0
posthog==3.5.0
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-doctr
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3.post1
pywin32==306
PyYAML==6.0
pyzmq==26.0.3
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
# scrapegraphai==0.9.0b7
selectolax==0.3.24
# selenium==4.15.2
sentence-transformers==2.4.0
serial==0.0.97
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.15.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
# spacy==3.7.2
# spacy-legacy==3.0.12
# spacy-loggers==1.0.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
sympy==1.12
tbb==2021.13.0
tenacity==8.2.3
tf2onnx == 1.8.4
thinc==8.2.3
threadpoolctl==3.3.0
tiktoken==0.5.2
tokenizers==0.15.2
toml==0.10.2
torch == 2.5.1
torchvision == 0.20.1
# torchaudio
tornado==6.4.1
tqdm==4.66.3
traitlets==5.14.3
transformers==4.38.1
trio==0.23.1
trio-websocket==0.11.1
# typer
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.9.0
urllib3 == 2.3.0
uvicorn == 0.32.0
virtualenv==20.25.1
wasabi==1.1.2
watchdog==6.0.0
watchfiles==0.22.0
wcwidth==0.2.13
# weasel==0.3.4
websocket-client==1.8.0
websockets==12.0
Werkzeug==1.0.1
wrapt==1.16.0
wsproto==1.2.0
yahoo-search-py==0.3
yarl==1.9.4
zipp==3.19.2
zope.interface==6.0
